datanode/requirements_datanode.txt
grpcio
grpcio-tools
requests
Flask